requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==5.0.2
django-jazzmin==2.6.0
psycopg2==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
